Feature/separate versions (#9)
* Create a common parent and common legacy project
