--- src/main/java/tds/common/ValidationError.java
@@ -1,32 +1,0 @@
-package tds.common;
-
-/**
- * Generic Error object
- */
-public class ValidationError {
-    private final String code;
-    private final String message;
-
-    /**
-     * @param code error code for the error type
-     * @param message error message to describe why error occurred
-     */
-    public ValidationError(String code, String message) {
-        this.code = code;
-        this.message = message;
-    }
-
-    /**
-     * @return error code
-     */
-    public String getCode() {
-        return code;
-    }
-
-    /**
-     * @return description why error occurred
-     */
-    public String getMessage() {
-        return message;
-    }
-}
